project configuration cleanup, warning fixes

+++ trunk/src/gb/gbMemory.cpp
@@ -217,7 +217,7 @@ void memoryUpdateMBC3Clock()
   time_t diff = now - gbDataMBC3.mapperLastTime;
   if(diff > 0) {
     // update the clock according to the last update time
-    gbDataMBC3.mapperSeconds += diff % 60;
+    gbDataMBC3.mapperSeconds += (int)(diff % 60);
     if(gbDataMBC3.mapperSeconds > 59) {
       gbDataMBC3.mapperSeconds -= 60;
       gbDataMBC3.mapperMinutes++;
@@ -225,7 +225,7 @@ void memoryUpdateMBC3Clock()
 
     diff /= 60;
 
-    gbDataMBC3.mapperMinutes += diff % 60;
+    gbDataMBC3.mapperMinutes += (int)(diff % 60);
     if(gbDataMBC3.mapperMinutes > 60) {
       gbDataMBC3.mapperMinutes -= 60;
       gbDataMBC3.mapperHours++;
@@ -233,14 +233,14 @@ void memoryUpdateMBC3Clock()
 
     diff /= 60;
 
-    gbDataMBC3.mapperHours += diff % 24;
+    gbDataMBC3.mapperHours += (int)(diff % 24);
     if(gbDataMBC3.mapperHours > 24) {
       gbDataMBC3.mapperHours -= 24;
       gbDataMBC3.mapperDays++;
     }
     diff /= 24;
 
-    gbDataMBC3.mapperDays += diff;
+    gbDataMBC3.mapperDays += (int)diff;
     if(gbDataMBC3.mapperDays > 255) {
       if(gbDataMBC3.mapperDays > 511) {
         gbDataMBC3.mapperDays %= 512;
